Update digiboard_change_wallpaper_to_windoge_10_and_start_rickroll_video.ino
This Digispark script works fine on windows 10 and windows 7 with a net runtime of 15 seconds as it needs to configure things before running the script it would take upto 5 seconds to pre-configure and then 10 seconds to run the script. I have tested this is windows 10 64-bit [8GB DDR3 ram core i3 8th gen on HDD] and windows 10 32-bit [4gb DDR4 ram core i3 8th gen on HDD] it works perfectly on both. But on windows 7 it might not change the background and needs to increase the delay between two commands.
Disclaimer :  This script only run in windows as it uses run program which is in-build in windows. You can make the same script for linux by using terminal instead of run program. 
The code in this file is been written by me and you are free to use it as you want. you might want to increase the delay if you want to use this script for low-end PC or laptop.
If you find any bug or want to make this script more efficient then please reach out to me on Instagram at @wraient.911 , your help will be appreciated.